Remove migrations feature & related docs.
