+++ protostar/commands/__init__.py
@@ -8,7 +8,6 @@ from .format_command import FormatCommand
 from .init import InitCommand
 from .install import InstallCommand
 from .invoke_command import InvokeCommand
-from .migrate_command import MigrateCommand
 from .migrate_configuration_file_command import MigrateConfigurationFileCommand
 from .remove import RemoveCommand, removal_exceptions
 from .test import TestCommand
