--- protostar/migrator/__init__.py
@@ -1,4 +1,0 @@
-from protostar.migrator.migrator import Migrator
-from protostar.migrator.migrator_execution_environment import (
-    MigratorExecutionEnvironment,
-)
