--- protostar/migrator/cheatcodes/__init__.py
@@ -1,1 +1,0 @@
-from .network_config import CheatcodeNetworkConfig, ValidatedCheatcodeNetworkConfig
