--- protostar/migrator/output_directory.py
@@ -1,28 +1,0 @@
-from contextlib import contextmanager
-from pathlib import Path
-
-from protostar.protostar_exception import ProtostarException
-
-
-@contextmanager
-def create_output_directory(path: Path):
-    if path.exists():
-        raise DirectoryExistsException(path)
-
-    try:
-        path.mkdir()
-        yield path
-    finally:
-        is_directory_empty = not any(path.iterdir())
-        if is_directory_empty:
-            path.rmdir()
-
-
-class DirectoryExistsException(ProtostarException):
-    def __init__(self, dir_path: Path) -> None:
-        super().__init__(
-            (
-                f"Directory ({dir_path}) exists.\n"
-                "Please delete, rename or relocate the directory."
-            )
-        )
