--- tests/e2e/test_migrating/box_contract.cairo
@@ -1,19 +1,0 @@
-%lang starknet
-from starkware.cairo.common.math import assert_nn
-from starkware.cairo.common.cairo_builtins import HashBuiltin
-
-@storage_var
-func box_value() -> (res: felt) {
-}
-
-@external
-func put{syscall_ptr: felt*, pedersen_ptr: HashBuiltin*, range_check_ptr}(value: felt) {
-    box_value.write(value);
-    return ();
-}
-
-@view
-func get{syscall_ptr: felt*, pedersen_ptr: HashBuiltin*, range_check_ptr}() -> (res: felt) {
-    let (res) = box_value.read();
-    return (res,);
-}
