--- tests/e2e/test_migrating/migration_up_down.cairo
@@ -1,18 +1,0 @@
-%lang starknet
-
-@external
-func up() {
-    %{
-        declaration = declare("./build/main.json")
-        deploy_contract(declaration.class_hash)
-    %}
-
-    return ();
-}
-
-@external
-func down() {
-    %{ declare("./build/main.json") %}
-
-    return ();
-}
