+++ website/docs/tutorials/07-testing/02-cheatcodes/declare.md
@@ -1,11 +1,7 @@
 # `declare`
 
 ```python
-def declare(
-    contract: str,
-    *,
-    config: Optional[CheatcodeNetworkConfig] = None
-) -> DeclaredContract:
+def declare(contract: str) -> DeclaredContract:
 
 class DeclaredContract:
     class_hash: int
@@ -13,4 +9,3 @@ class DeclaredContract:
 Declares contract given a path relative to a Protostar project root.
 
 - `contract` — A path to the contract's main `.cairo` source file.
-- `config` — A keyword only argument kept for compatibility with the migration [declare cheatcode](../../09-migrations/declare.md). See related documentation for more information.
