+++ website/docs/tutorials/07-testing/02-cheatcodes/deploy-contract.md
@@ -4,8 +4,6 @@
 def deploy_contract(
     contract: str,
     constructor_calldata: Optional[Union[List[int], Dict]] = None,
-    *,
-    config: Optional[CheatcodeNetworkConfig] = None
 ) -> DeployedContact:
 
 class DeployedContract:
@@ -14,7 +12,6 @@ class DeployedContract:
 Deploys a contract given a path relative to a Protostar project root. The section [Deploying contracts from tests](../01-deploying-contracts.md) demonstrates a usage of this cheatcode.
 
 - `contract` — A path to the contract's main `.cairo` source file.
-- `config` — A keyword only argument kept for compatibility with the migration [deploy_contract cheatcode](../../09-migrations/deploy-contract.md). See related documentation for more information.
 
 :::warning
 Deploying a contract is a slow operation. If it's possible try using this cheatcode in the [`__setup__` hook](../README.md#setup-hooks).
