+++ website/docs/tutorials/09-migrations/_category_.yaml
@@ -1,0 +1,1 @@
+label: 'Migrations'
