--- website/docs/tutorials/09-migrations/network-config.md
@@ -1,21 +1,0 @@
-# Network configuration
-```python
-class CheatcodeNetworkConfig(TypedDict):
-    wait_for_acceptance: bool
-    max_fee: NotRequired[Wei | "auto"]
-    
-```
-
-- `max_fee` - The maximum fee that the sender is willing to pay for the transaction. Required for transactions V1.
-- `wait_for_acceptance` - Wait synchronously for transaction to be accepted on chain.
-
-`CheatcodeNetworkConfig` stores configuration used in `declare` and `deploy_contract` cheatcodes.
-
-It has no effect in the test environment, however it is kept for compatibility.
-
-#### Default configuration:
-```python
-CheatcodeNetworkConfig(
-    wait_for_acceptance=False
-)
-```
